Erase all my Class Controllers and use the clasic one wrote cleaner and scalable code
