--- backend/config/db.js
@@ -1,11 +1,0 @@
-import mongoose from "mongoose";
-
-export const connectDB = async () => {
-  try {
-    await mongoose.connect(process.env.MONGO_URI);
-    console.log("MongoDB connected successfully");
-  } catch (error) {
-    console.log("MongoDB connection failed");
-    process.exit(1);
-  }
-};
