--- backend/middlewares/admin.js
@@ -1,44 +1,0 @@
-import jwt from "jsonwebtoken";
-import User from "../models/User.js";
-import dotenv from "dotenv";
-
-dotenv.config();
-
-export const protect = async (req, res, next) => {
-  let token = req.cookies.access_token;
-  if (
-    req.headers.authorization &&
-    req.headers.authorization.startsWith("Bearer")
-  ) {
-    token = req.headers.authorization.split(" ")[1];
-  }
-
-  if (!token) {
-    return res.status(401).json({ message: "Not authorized" });
-  }
-
-  try {
-    const decoded = jwt.verify(token, process.env.SECRET);
-
-    req.user = await User.findById(decoded.userId).select("-password");
-
-    next();
-  } catch (error) {
-    res.status(401).json({ message: `Token failed: ${error}` });
-  }
-};
-
-// Middleware to allow only admins
-export const adminOnly = (req, res, next) => {
-  try {
-    if (!req.user) {
-      return res.status(401).json({ message: "User not authenticated" });
-    }
-
-    if (req.user.role !== "admin") {
-      return res.status(403).json({ message: "Admins only!" });
-    }
-  } catch (error) {}
-
-  next();
-};
