--- backend/server.js
@@ -1,27 +1,0 @@
-import express from "express";
-import dotenv from "dotenv";
-import cookieParser from "cookie-parser";
-import { connectDB } from "./config/db.js";
-import productRoutes from "./routes/productRoutes.js";
-import userRoutes from "./routes/userRoutes.js";
-dotenv.config();
-
-const app = express();
-// Middleware
-app.use(express.json());
-app.use(cookieParser());
-// Routes
-app.use("/api/products", productRoutes);
-app.use("/api/users", userRoutes);
-
-// Connect to MongoDB
-connectDB()
-  .then(() => {
-    app.listen(process.env.PORT, () => {
-      console.log(`Server is running on http://localhost:${process.env.PORT}`);
-    });
-  })
-  .catch((error) => {
-    console.error("MongoDB connection failed", error);
-    process.exit(1);
-  });
