+++ backend/src/constants/http.ts
@@ -1,0 +1,10 @@
+export const OK = 200;
+export const CREATED = 201;
+export const BAD_REQUEST = 400;
+export const UNAUTHORIZED = 401;
+export const FORBIDDEN = 403;
+export const NOT_FOUND = 404;
+export const CONFLICT = 409;
+export const UNPROCESSABLE_CONTENT = 422;
+export const TOO_MANY_REQUEST = 429;
+export const INTERNAL_SERVER_ERROR = 500;
