+++ backend/src/models/Reviews.ts
@@ -1,0 +1,27 @@
+import mongoose from "mongoose";
+
+const reviewSchema = new mongoose.Schema(
+  {
+    rating: {
+      type: Number,
+      required: true,
+      min: 1,
+      max: 5,
+    },
+    comment: {
+      type: String,
+      required: true,
+    },
+    reviewerName: {
+      type: String,
+      ref: "User",
+    },
+  },
+  {
+    timestamps: true,
+  }
+);
+
+const Review = mongoose.model("Review", reviewSchema);
+
+export default Review;
