+++ backend/src/utils/catchErrors.ts
@@ -1,0 +1,19 @@
+import { NextFunction, Request, Response } from "express";
+
+type AsyncController = (
+  req: Request,
+  res: Response,
+  next: NextFunction
+) => Promise<any>;
+
+const catchErrors =
+  (controller: AsyncController): AsyncController =>
+  async (req, res, next) => {
+    try {
+      await controller(req, res, next);
+    } catch (error) {
+      next(error);
+    }
+  };
+
+export default catchErrors;
